Refactor game logic to improve card management and introduce new game stages
